move downstream to its own separate module
